use improved fonts with next/font
Switch to using next/font instead of pure CSS font. next/font improves
the display of custom fonts when they are being loaded. If you use pure
CSS font and the font is being loaded, it may not appear until it's
fully loaded. With next/font, the font will be preloaded using the
default font.

__ Correct me if I'm wrong 😁 __

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>