add a <UIDInput/> component
This component will be implemented in the main index page to display an
input UID box. This input box will be used to fetch a player's profile
using their UID.

Altough, the `UIDInput.module.css` currently not in used because it's
still in development, this file will be used in the future.

Signed-off-by: Muhammad Rizki <kiizuha@gnuweeb.org>